Only push the user details view automatically onto the feedback list stack once at max
If the required manual user data is missing and the feedback list view appears the user detail view is automatically pushed. Now if the user cancels and goes back to the feedback list, it gets pushed again and so the user can never exit. Instead push this at maximum once. When providing feedback it will ask the missing user details again anyway.