More changes to the behavior of first forced user data and compose UI presentation
This is a change to previous behavior! Now the SDK by default shows the user data UI and subsequent compose UI always modally to ensure proper functionality on all scenarios the feedback viewer can be presented.

The new property `showFirstRequiredPresentationModal` in `BITFeedbackManager` can be set to `NO` if the previous behavior of pushing the UI and Compose view onto the stack, when those are being forced to be shown on the first presentation of the feedback UI. Especially if you present the UI in a popover, you should NOT change this property!