Use only one activity view controller per UIActivity
On iPad under iOS 6, calling `activityDidFinish:` internally calls -[BITFeedbackActivity activityViewController]. We must therefore keep a reference to the previously created activity view controller else the activity view controller can not be dismissed since a new instance is returned for dismissal. Under iOS 7, -[BITFeedbackActivity activityViewController] is not called for dismissal.

On iPhone, the -[BITFeedbackActivity activityViewController] method is not called for dismissal.