Add workaround for iOS 8 update process issue
In iOS 8 the app is not moved to background once the users accepts the iOS install alert button. Without this, the install process doesn't start. So for now we exit the app once that alert disappears.

Important: The iOS dialog offers the user to deny installation, we can't find out which button was tapped, so we assume the user agreed.